# rework Photoneo YCoCg(-R) for SIMD autovectorization