Related to #15, #13

Photoneo `YCoCg(-R)` subtly differs from VESA DSC  `YCoCg-R`
- https://github.com/photoneo-3d/photoneo-cpp-examples/issues/4
  - extra check for y==0 zeroying out pixel to protect black pixels in 4:2:0 subsampling
  - potentially small differences in LSbs, some of which are already being corrected
    - when scaling to 8 bit RGB this difference is not visible anyway

- rework our implementation so that it autovectorizes 
- add commented out CMake line to enable autovectoratization
  - do not enable native optimization with vectorization for generic build
    - code generated this way may not be portable to different architectures

------------------------------------

Powersafe

| No optimization | Optimized | Autovectorized without y==0 check | Autovectorized with y==0 check | Autovectorized with optimized  y==0 check |
|----------------------|------------------|------------------------------------------------|--------------------------------------------|--------------|
| up to 30 ms      |  up to 11 ms|  2-3 ms                                              | Up to 7-8 ms                               |    2-3 ms     |

-----------------------------

Performance (for reference, the code would not typically run this way)

| No optimization | Optimized | Autovectorized without y==0 check | Autovectorized with y==0 check |  Autovectorized with optimized  y==0 check |
|----------------------|------------------|------------------------------------------------|--------------------------------------------|--------|
| 7 ms      | 3 ms  |  1 ms                                                                           |  2 ms   |  1 ms |